MWPW-167747 Optimize acrobat action binder with separated upload logic (#260)
- Refactor action binder to split upload functions per guest/user flows
- Create new `upload-handler` to encapsulate file upload logic
- Optimize `action-binder` for performance and reduce file size
- Allows for disabling of signed-in flows for future verbs without the
need to load any upload logic

Resolves: [MWPW-167747](https://jira.corp.adobe.com/browse/MWPW-167747)

**Test URLs:**

Compress:
- Before:
https://stage--dc--adobecom.hlx.page/acrobat/online/compress-pdf?unitylbs=stage&martech=off
- After:
https://stage--dc--adobecom.hlx.page/acrobat/online/compress-pdf?unitylbs=MWPW-167747-ab-update&martech=off

F&S:
- Before:
https://stage--dc--adobecom.hlx.page/acrobat/online/sign-pdf?unitylbs=stage&martech=off
- After:
https://stage--dc--adobecom.hlx.page/acrobat/online/sign-pdf?unitylbs=MWPW-167747-ab-update&martech=off
